Implement no op Billing
