+++ src/main/java/io/stargate/sgv2/jsonapi/api/request/RequestContext.java
@@ -16,6 +16,7 @@ import io.stargate.sgv2.jsonapi.config.feature.ApiFeatures;
 import io.stargate.sgv2.jsonapi.config.feature.FeaturesConfig;
 import io.stargate.sgv2.jsonapi.logging.LoggingMDCContext;
 import io.stargate.sgv2.jsonapi.service.provider.Billing;
+import io.stargate.sgv2.jsonapi.service.provider.LoggingBilling;
 import io.stargate.sgv2.jsonapi.service.schema.SchemaRegistry;
 import io.vertx.ext.web.RoutingContext;
 import jakarta.enterprise.context.RequestScoped;
@@ -222,7 +223,7 @@ public class RequestContext implements LoggingMDCContext {
     if (billing == null) {
       synchronized (this) {
         if (billing == null) {
-          billing = new Billing(commandConfig.get(BillingConfig.class), apiFeatures());
+          billing = new LoggingBilling(commandConfig.get(BillingConfig.class), apiFeatures());
         }
       }
     }
